debug

+++ openrlhf/trainer/ray/ppo_actor.py
@@ -365,8 +365,8 @@ class ActorPPOTrainer(ABC):
         print("output_logits", output["logits"].shape) # batch-size, sequence-len, 1, vocab
         print("output_full_logits", output_full["logits"].shape) # batch-size, sequence-len, 1, vocab
         with torch.no_grad():
-            output_logits = output["logits"].squeeze(2)[target_index: target_index + 1, :-1, :][action_mask[target_index: target_index + 1] == 1]
-        output_full_logits = output_full["logits"].squeeze(2)[:, :-1, :][action_mask_full == 1]
+            output_logits = output["logits"].squeeze(2)[target_index: target_index + 1, :-1, :][action_mask[target_index: target_index + 1] == 1].to(torch.bfloat16)
+        output_full_logits = output_full["logits"].squeeze(2)[:, :-1, :][action_mask_full == 1].to(torch.bfloat16)
         print("output_logits", output_logits.shape) # 1, response-len, vocab
         print("output_full_logits", output_full_logits.shape) # 1, response-len, vocab
         kd_loss = F.kl_div(
